backend/requirements.txt
asgiref==3.9.1
boto3==1.40.13
botocore==1.40.13
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.5
django-pagination==1.0.7
django-render-block==0.11
django-sendgrid-v5==1.3.0
django-storages==1.14.6
django-templated-email==3.1.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
MarkupSafe==3.0.3
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
requests==2.32.5
s3transfer==0.13.1
sendgrid==6.12.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.9.0
